Use s390 return-in-memory behavior for now, add some defs to zos.h
Reenable s390_return_in_memory, so now we will use the same logic for
deciding what return values go in memory as is used in z/Linux, which
actually turns out to be very close to what we want. The points of
contention are vector types >= 128-bits wide and floating-point types.

For the moment, we will now return vectors in vector registers if they
can fit, and floating point values however s390 does it.

This fixes an internal compiler error that occured whenever trying to
compile a function returning a long double when the -mlong-double-128
option was in effect.

Also add some defines to zos.h that seem beneficial, and rewrite the
register allocation order a bit.