Rewrite the prologue and epilogue
Fix bugs that prevented saves in leaf functions. Allow the backend to
use fprs to save regs in leaf functions. Remove the unnecessary use
of r14 in the prologue, so it now clobbers no call-saved regs. Don't
create a DSA in functions that don't require one, significantly
reducing the size of very small functions.