Implement consistent DWARF generation for z/OS
We implement DWARF by defining the CFA to be the value of the frame
pointer (r13) on entry to the function, which allows us to note all of
our register saves with a regular cfi_offset. To get this to work with
potentially discontinuous stack model, we have to use a
DW_CFA_val_expression and a DW_CFA_def_cfa_expression to specify where
we store the old CFA address and the old r13 value respectively when we
change the value of R13 in the prologue. This also means that the
floating point register saves can only be defined with
DW_CFA_val_expression rules, which we consider to be acceptable.

INCU-424