Change __builtin_dwarf_sp_column() to return FP
The __builtin_dwarf_sp_column() builtin is meant to return dwarf column
corresponding to a register that should be the CFA at entrance to any
given function. Right now, we return r11, which we occasionally turn
into a temporary stack pointer, but has no meaning on entrance to the
function.

Despite the bultin's name, it seems like the correct behavior would be
to return the frame pointer (r13) dwarf column for z/OS.
__builtin_dwarf_sp_column only seems to be used inside the unwinder,
and changing it in such a way would make it work in that context, so
there doesn't seem to be a reason not to change it in such a way.

This problem seems to be blocking exception handling.

INCU-434